+++ src/WebJobs.Script/WebJobs.Script.csproj
@@ -44,10 +44,10 @@
     <PackageReference Include="Microsoft.Azure.Functions.NodeJsWorker" Version="2.1.1" />
     <PackageReference Include="Microsoft.Azure.Functions.PowerShellWorker.PS6" Version="3.0.630" />
     <PackageReference Include="Microsoft.Azure.Functions.PowerShellWorker.PS7" Version="3.0.738" />
-    <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.1.0-11842" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="4.0.3" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Http" Version="3.1.0" />
-    <PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="4.1.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="4.1.0-11842" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Logging.ApplicationInsights" Version="3.0.28" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Script.Abstractions" Version="1.0.2-preview" />
     <PackageReference Include="Microsoft.Build" Version="15.8.166" />
