+++ test/WebJobs.Script.Tests.Integration/WebJobs.Script.Tests.Integration.csproj
@@ -39,9 +39,9 @@
     <PackageReference Include="Microsoft.Azure.DocumentDB.Core" Version="2.10.0" />
     <PackageReference Include="Microsoft.Azure.EventHubs" Version="2.1.0" />
     <PackageReference Include="Microsoft.Azure.Functions.NodeJsWorker" Version="2.1.1" />
-    <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.1.0-11842" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Storage" Version="4.0.4" />
-    <PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="4.1.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="4.1.0-11842" />
     <PackageReference Include="Microsoft.Azure.Functions.JavaWorker" Version="1.8.2-SNAPSHOT" />
     <PackageReference Include="Microsoft.Azure.Mobile.Client" Version="4.0.2" />
     <PackageReference Include="Microsoft.Azure.ServiceBus" Version="3.1.0" />
@@ -104,13 +104,7 @@
     <None Include="ServiceBus\ServiceBusTriggerNodeEndToEndTests.cs" />
     <None Include="Twilio\TwilioEndToEndTestsBase.cs" />
   </ItemGroup>
-  
-  <ItemGroup>
-    <Content Update="appsettings.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-   
+
   <Import Project="..\..\build\GrpcTestFix.targets" />
   <Import Project="..\WebJobs.Script.Tests.Shared\WebJobs.Script.Tests.Shared.projitems" Label="Shared" />
